+++ certbot-apache/certbot_apache/_internal/entrypoint.py
@@ -49,7 +49,7 @@ def get_configurator() -> Type[configurator.ApacheConfigurator]:
 
     # Special case for older Fedora versions
     min_version = util.LooseVersion('29')
-    if os_name == 'fedora' and util.LooseVersion(os_version) < min_version:
+    if os_name == 'fedora' and util.LooseVersion(os_version).try_risky_less(min_version):
         os_name = 'fedora_old'
 
     try: